Clean some compilation warnings.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/crossfire@3747 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ common/button.c
@@ -252,7 +252,7 @@ void animate_turning(object *op) /* only one part objects */
 }
 
 #define ARCH_SACRIFICE(xyz) ((xyz)->slaying)
-#define NROF_SACRIFICE(xyz) ((xyz)->stats.food)
+#define NROF_SACRIFICE(xyz) ((uint32)(xyz)->stats.food)
 
 /* Returns true if the sacrifice meets the needs of the altar.
  *
